Use build script to generate const.rs file
