+++ gen/genmult/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "libsecp256k1-gen-genmult"
-version = "0.0.0"
+version = "0.4.0"
 authors = ["Wei Tang <hi@that.world>"]
 edition = "2018"
